+++ CHANGELOG.md
@@ -8,6 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Added
   - Experimental support for downloading F2D derivatives
+- Changed
+  - Moved to viewer version 7.9.* (currently with a small hack
+  to disable writing to `document.cookie` before it's fixed on the viewer side)
 
 ## [1.4.8] - 2019-12-11